rpi/requirements.txt
python-socketio[asyncio-client]
asyncio
numpy
pyserial
